backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pillow==10.3.0
qrcode[pil]==7.4.2
boto3==1.34.95
aiofiles==23.2.1
httpx==0.27.0
redis==5.0.4
celery==5.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
reportlab==4.2.0
openpyxl==3.1.2
pandas==2.2.2
websockets==12.0
pytest==8.2.0
pytest-asyncio==0.23.6
